separation of concerns by moving api and auth routes to separate files, moved passport config to it's own file and directory
